requirements.txt
PyYAML==6.0.1
regex==2024.4.16
tinycss2==1.2.1
